requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.19
numpy>=1.24.0
